requirements.txt
aiogram==3.10.0
python-dotenv
gigachat
openai-whisper
pydub
graphviz
fpdf2